+++ src/components/table-ajax/__spec__.js
@@ -11,6 +11,7 @@ describe('TableAjax', () => {
 
     instance = TestUtils.renderIntoDocument(
       <TableAjax
+        actions={ { foo:{}, bar:{} } }
         className="foo"
         path='/test'
         onChange={ spy }
@@ -21,6 +22,7 @@ describe('TableAjax', () => {
 
     customInstance = TestUtils.renderIntoDocument(
       <TableAjax
+        actions={ { foo:{}, bar:{} } }
         className="foo"
         path='/test'
         onChange={ spy }
@@ -33,6 +35,7 @@ describe('TableAjax', () => {
 
     pageSizeInstance = TestUtils.renderIntoDocument(
       <TableAjax
+        actions={ { foo:{}, bar:{} } }
         className="foo"
         path='/test'
         onChange={ spy }