+++ src/components/table/__spec__.js
@@ -19,7 +19,7 @@ describe('Table', () => {
     );
 
     instance = TestUtils.renderIntoDocument(
-      <Table className="foo">
+      <Table className="foo" actions={{ foo:{}, bar:{} }}>
         { row }
       </Table>
     );
@@ -38,7 +38,7 @@ describe('Table', () => {
     );
 
     instanceSortable = TestUtils.renderIntoDocument(
-      <Table className='bar' onChange={ spy }>
+      <Table className='bar' onChange={ spy } actions={ { foo:{}, bar:{} } }>
         <TableRow key='header'>
           <TableHeader sortable={true} name='name'>
             Names
@@ -173,7 +173,7 @@ describe('Table', () => {
 
       beforeEach(() => {
         spy = jasmine.createSpy();
-        instance = TestUtils.renderIntoDocument(<Table onSelect={ spy } selectable={ true }><TableRow uniqueID="foo" /></Table>);
+        instance = TestUtils.renderIntoDocument(<Table onSelect={ spy } selectable={ true } actions={ { foo:{}, bar:{} } }><TableRow uniqueID="foo" /></Table>);
         row = TestUtils.findRenderedComponentWithType(instance, TableRow);
         spyOn(row, 'setState');
       });
@@ -195,7 +195,7 @@ describe('Table', () => {
 
       describe('if row is to be selected but is selectAll', () => {
         it('does not add the row', () => {
-          instance = TestUtils.renderIntoDocument(<Table onSelect={ spy } selectable={ true }><TableRow uniqueID="foo" selectAll={ true } /></Table>);
+          instance = TestUtils.renderIntoDocument(<Table onSelect={ spy } selectable={ true }  actions={ { foo:{}, bar:{} } }><TableRow uniqueID="foo" selectAll={ true } /></Table>);
           row = TestUtils.findRenderedComponentWithType(instance, TableRow);
           instance.selectRow('foo', row, true);
           expect(instance.selectedRows["foo"]).toBe(undefined);
@@ -220,7 +220,7 @@ describe('Table', () => {
 
     describe('if highlight row', () => {
       beforeEach(() => {
-        instance = TestUtils.renderIntoDocument(<Table highlightable={ true }><TableRow uniqueID="foo" /></Table>);
+        instance = TestUtils.renderIntoDocument(<Table highlightable={ true }  actions={ { foo:{}, bar:{} } }><TableRow uniqueID="foo" /></Table>);
         row = TestUtils.findRenderedComponentWithType(instance, TableRow);
         spyOn(row, 'setState');
       });
@@ -285,7 +285,7 @@ describe('Table', () => {
     describe('if there is an onHighlight callback', () => {
       it('calls the callback', () => {
         let spy = jasmine.createSpy();
-        instance = TestUtils.renderIntoDocument(<Table highlightable={ true } onHighlight={ spy }><TableRow uniqueID="foo" /></Table>);
+        instance = TestUtils.renderIntoDocument(<Table highlightable={ true } onHighlight={ spy }  actions={ { foo:{}, bar:{} } }><TableRow uniqueID="foo" /></Table>);
         row = TestUtils.findRenderedComponentWithType(instance, TableRow);
         instance.highlightRow('foo', row, true);
         expect(spy).toHaveBeenCalledWith('foo', true, row);
@@ -340,7 +340,7 @@ describe('Table', () => {
       it('calls the callback', () => {
         let spy = jasmine.createSpy();
         instance = TestUtils.renderIntoDocument(
-          <Table onSelect={ spy } />
+          <Table onSelect={ spy }  actions={ { foo:{}, bar:{} } } />
         );
         instance.rows = {};
         row = { state: {}, setState: () => {} };
@@ -353,7 +353,7 @@ describe('Table', () => {
       it('calls setState', () => {
         let spy = jasmine.createSpy();
         instance = TestUtils.renderIntoDocument(
-          <Table />
+          <Table  actions={ { foo:{}, bar:{} } } />
         );
         instance.actionToolbarComponent = {
           setState: spy
@@ -410,7 +410,7 @@ describe('Table', () => {
     beforeEach(() => {
       data = Immutable.fromJS({ foo: "bar" });
       instance = TestUtils.renderIntoDocument(
-        <Table filter={ data }></Table>
+        <Table filter={ data }  actions={ { foo:{}, bar:{} } }></Table>
       );
       spyOn(instance, 'emitOnChangeCallback');
     });
@@ -439,7 +439,7 @@ describe('Table', () => {
     describe('when highlightable is disabled', () => {
       it('resets the highlighted row', () => {
         instance = TestUtils.renderIntoDocument(
-          <Table highlightable={ true } />
+          <Table highlightable={ true }  actions={ { foo:{}, bar:{} } } />
         );
         spyOn(instance, 'resetHighlightedRow');
         instance.componentWillReceiveProps({ highlightable: false });
@@ -450,7 +450,7 @@ describe('Table', () => {
     describe('when selectable is disabled', () => {
       it('resets the selectable rows', () => {
         instance = TestUtils.renderIntoDocument(
-          <Table selectable={ true } />
+          <Table selectable={ true }  actions={ { foo:{}, bar:{} } } />
         );
         instance.rows = {
           foo: { props: { uniqueID: 'foo' } }
@@ -545,7 +545,7 @@ describe('Table', () => {
     describe('when shrink is enabled', () => {
       beforeEach(() => {
         instance = TestUtils.renderIntoDocument(
-          <Table shrink={ true }></Table>
+          <Table shrink={ true }  actions={ { foo:{}, bar:{} } }></Table>
         );
       });
 
@@ -746,7 +746,7 @@ describe('Table', () => {
     describe('when pageSize is passed', () => {
       it('returns the prop pageSize', () => {
         instance = TestUtils.renderIntoDocument(
-          <Table path='/test' pageSize='123' >
+          <Table path='/test' pageSize='123'  actions={ { foo:{}, bar:{} } }>
           </Table>
         );
         expect(instance.defaultPageSize).toEqual('123')
@@ -762,7 +762,7 @@ describe('Table', () => {
         ]);
 
         instance = TestUtils.renderIntoDocument(
-          <Table pageSizeSelectionOptions={ options } >
+          <Table pageSizeSelectionOptions={ options }  actions={ { foo:{}, bar:{} } }>
           </Table>
         );
         expect(instance.defaultPageSize).toEqual('1')
@@ -780,7 +780,7 @@ describe('Table', () => {
     describe('when thead is not provided', () => {
       it('returns the the correct markup', () => {
         instance = TestUtils.renderIntoDocument(
-          <Table path='/test'>
+          <Table path='/test'  actions={ { foo:{}, bar:{} } }>
           </Table>
         );
         let parent = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'thead')[0];
@@ -798,7 +798,7 @@ describe('Table', () => {
           </TableRow>
         );
         instance = TestUtils.renderIntoDocument(
-          <Table path='/test' thead={header}>
+          <Table path='/test' thead={header}  actions={ { foo:{}, bar:{} } }>
           </Table>
         );
         let parent = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'thead')[0];
@@ -818,7 +818,7 @@ describe('Table', () => {
     });
 
     it('returns custom tbody when setting tbody prop to false', () => {
-      instance = TestUtils.renderIntoDocument(<Table tbody={ false }><tbody className="custom"><tr><td></td></tr></tbody></Table>);
+      instance = TestUtils.renderIntoDocument(<Table tbody={ false }  actions={ { foo:{}, bar:{} } }><tbody className="custom"><tr><td></td></tr></tbody></Table>);
       expect(instance.tbody.props.className).toEqual("custom");
     });
   });